pyproject.toml
[project]
name = "auto-wod-booking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "pdbpp>=0.10.3",
    "python-dotenv>=1.0.1",
    "selenium>=4.28.1",
]

[dependency-groups]
dev = ["black>=25.1.0", "ruff>=0.9.4"]
